chore: added basic stuff
