+++ packages/sidecar/src/mcp/utils/index.ts
@@ -1,0 +1,2 @@
+export * from "./errors.js";
+export * from "./logs.js";
